Refactor structure of files to be like original
