+++ client/src/components/achievements/Achievement.js
@@ -17,9 +17,9 @@ import { IconContext } from 'react-icons';
 import { AiFillDelete, AiFillEdit } from 'react-icons/ai';
 import { useDispatch } from 'react-redux';
 import { PropTypes } from 'prop-types';
-import './css/achievementForm.css';
+import './form/achievementForm.css';
 import { deleteAchievement } from '../../actions/achievements';
-import AchievementForm from './achievementsComponents/AchievementForm';
+import AchievementForm from './form/AchievementForm';
 
 const Achievement = ({ achievement }) => {
   const dispatch = useDispatch();
