--- client/src/components/achievements/achievementsComponents/AddButton.jsx
@@ -1,18 +1,0 @@
-/**
- * File exports add button
- */
-import React from 'react';
-import { Button } from '@material-ui/core';
-
-const AddButton = () => (
-  <Button
-    variant="contained"
-    size="large"
-    color="primary"
-  >
-    kfmvkfvr
-  </Button>
-
-);
-
-export default AddButton;
